+++ web/src/main/java/com/epam/esm/dto/TagDto.java
@@ -4,19 +4,19 @@ import java.util.Objects;
 import java.util.StringJoiner;
 
 public class TagDto {
-    private int id;
+    private long id;
     private String name;
 
-    public TagDto(int id, String name) {
+    public TagDto(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
